Code to find minimum distance between 2 states
The code written in C++, takes in input a file with the configuration of
the search space, start state, goal state. It then uses A* algorithm to
search the minimum path and writes the nodes of the minimum cost path to
a file 'Solution.txt'.  This is my solution to an assignment from
Artificial Intelligence Course (Monsoon'15, IIIT-Delhi).